remove some warning
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@5590 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/daemon/newpop3.c
@@ -303,7 +303,7 @@ void outfile(filename, linenum)
     char linebuf[10240];
     char *buf, *p;              /* KCN.99.09.01 */
     char *ptr;
-    size_t size;
+    off_t size;
     long left;
 
     if (linenum) {
